modifiquei add-anuncio.php, anuncios.class.php, meus-anuncios.php e criei os arquivos editar-anuncio.php, excluir-anuncio.php e categorias.class.php

+++ classificados/meus-anuncios.php
@@ -31,16 +31,16 @@ if(empty($_SESSION['cLogin'])) {
 		<tr>
 			<td>
 				<?php if(!empty($anuncio['url'])): ?>
-				<img src="assets/images/anuncios/<?php echo $anuncio['url']; ?>" height="50" border="0" />
+                    <img src="assets/images/anuncios/<?php echo $anuncio['url']; ?>" height="50" border="0" />
 				<?php else: ?>
-				<img src="assets/images/default.jpg" height="50" border="0" />
+                    <img src="assets/images/default.jpg" height="50" border="0" />
 				<?php endif; ?>
 			</td>
 			<td><?php echo $anuncio['titulo']; ?></td>
 			<td>R$ <?php echo number_format($anuncio['valor'], 2); ?></td>
 			<td>
 				<a href="editar-anuncio.php?id=<?php echo $anuncio['id']; ?>" class="btn btn-default">Editar</a>
-				<a href="excluir-anuncio.php?id=<?php echo $anuncio['id']; ?>" class="btn btn-danger">Excluir</a>
+				<a href="excluir-anuncio.php?id=<?php echo $anuncio['id']; ?>" class="btn btn-danger" onclick="return confirm('Tem certeza que deseja excluir este item?');">Excluir</a>
 			</td>
 		</tr>
 		<?php endforeach; ?>
